backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.33",
    "langfuse>=3.5.0",
    "langgraph>=0.6.7",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "uvicorn>=0.36.0",
    "httpx>=0.25.2",
    "python-socketio==5.10.0",
    
    # Authentication & Security
    "argon2-cffi>=23.1.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.1.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
